Code debug cleanup
After turning on pre-launch chloe we fixed any bugs to version 6.0.3. This meant adding console logs throughout the code to debug. To officially turn Chloe on we want these debug console log lines of code removed so this is not a version update just a clean up to the code to remove any missed lines of code used for debugging.